requirements.txt
python-binance==0.7.12
python-dotenv==0.19.2
